src/requirements.txt
pandas==0.22.0
numpy==1.14.0
Flask==0.12.3
Flask-Cors==3.0.3
psycopg2-binary==2.7.7
SQLAlchemy==1.3.0
jsonschema==3.2.0
PuLP==2.7.0
geopy 